+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # Changelog
-## Unreleased 
+## Unreleased
+- Add default `ORDER BY` primary key when no order is specified to ensure deterministic pagination results ([#73](https://github.com/aarondfrancis/fast-paginate/issues/73))